provide an option to use openssl to sign instead of ring
