+++ Cargo.toml
@@ -25,11 +25,13 @@ path = "src/lib.rs"
 [features]
 # This library was first created to support GCP oauth, if we add support for
 # other oauth providers this will most likely change to not have any default features
-default = ["gcp"]
+default = ["gcp", "sign-ring"]
 # Supports for GCP oauth2
 gcp = ["jwt", "url"]
-# Support for Json Web Tokens, ring is used for signing
-jwt = ["ring"]
+# Support for Json Web Tokens, ring is used for signing if enable feature `sign-ring`
+jwt = []
+sign-ring = ["ring"]
+sign-ssl = ["openssl"]
 # This enables features in chrono and ring that are necessary to use this library
 # in a wasm32 web (browser) context. If you are using wasm outside the browser
 # you will need to target wasm32-wasi for the requisite functionality (time and random)
@@ -38,6 +40,7 @@ wasm-web = ["ring/wasm32_unknown_unknown_js"]
 [dependencies]
 data-encoding = "2.4"
 http = "0.2"
+openssl = { version = "0.10", optional = true }
 ring = { version = "0.17", optional = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
